+++ src/anemoi/training/losses/logcosh.py
@@ -28,7 +28,6 @@ class WeightedLogCoshLoss(BaseWeightedLoss):
     def __init__(
         self,
         node_weights: torch.Tensor,
-        variable_scaling: torch.Tensor | None = None,
         ignore_nans: bool = False,
         **kwargs,
     ) -> None:
@@ -38,15 +37,12 @@ class WeightedLogCoshLoss(BaseWeightedLoss):
         ----------
         node_weights : torch.Tensor of shape (N, )
             Weight of each node in the loss function
-        variable_scaling : Optional[torch.Tensor], optional
-            precomputed, per-variable weights, by default None
         ignore_nans : bool, optional
             Allow nans in the loss and apply methods ignoring nans for measuring the loss, by default False
 
         """
         super().__init__(
             node_weights=node_weights,
-            variable_scaling=variable_scaling,
             ignore_nans=ignore_nans,
             **kwargs,
         )